+++ src/pages/Home.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Category, Hero, Trending } from "../components/Home";
+import { Category, Hero, Trending } from "../components/Home/index.js";
 
 export const Home = () => {
   return (
